[refactor] move command_buffer.py
