+++ device/__init__.py
@@ -1,1 +1,2 @@
 from device.device import Device
+from device.command_buffer import CommandBuffer
