+++ device/ssd.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 
-from command_buffer.command_buffer import CommandBuffer
 from device import Device
+from command_buffer import CommandBuffer
 import argparse
 import re
 
@@ -80,8 +80,10 @@ class SSD(Device):
             return
 
         if self.command_buffer:
+            print("in")
             self.command_buffer.add_command(cmd_type="W", lba=lba, value_or_size=value)
         else:
+            print("12233")
             data = self.nand.load()
             data[lba] = value
             self.nand.save(data)
