+++ tests/test_ssd.py
@@ -1,13 +1,17 @@
 import sys
+import os
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '../device')))
+
 import pytest
 import shutil
 
 from dataclasses import dataclass
 from pytest_mock import MockerFixture
 
-from command_buffer.command_buffer import CommandBuffer
 from device import ssd
 from device.ssd import SSD, INITIALIZED_DATA, ERROR, OUTPUT_DIR, NAND, OutputWriter, Validator
+from device.command_buffer import CommandBuffer
+
 
 
 @dataclass
@@ -29,7 +33,7 @@ def ssd_instance():
                )
 
     command_buffer.on_flush_callback = ssd1.flush
-
+    command_buffer.initialize_buffer()
     return ssd1
